+++ src/jdbc_41/org/firebirdsql/jdbc/FBResultSet.java
@@ -57,4 +57,10 @@ public class FBResultSet extends AbstractResultSet {
             throws SQLException {
         super(rowDescriptor, connection, rows, retrieveBlobs);
     }
+
+    public FBResultSet(RowDescriptor rowDescriptor, FBConnection connection, List<RowValue> rows,
+            FBObjectListener.ResultSetListener listener, boolean retrieveBlobs, boolean trimStrings)
+            throws SQLException {
+        super(rowDescriptor, connection, rows, listener, retrieveBlobs, trimStrings);
+    }
 }