+++ src/documentation/release_notes.md
@@ -12,7 +12,9 @@ The following has been changed or fixed since Jaybird 4.0.8
 - Fixed: Some methods of `FBClob` threw `NullPointerException` when attempting
   to use after implicit or explicit `free()` ([jaybird#719](https://github.com/FirebirdSQL/jaybird/issues/719))
 - Improvement: If the default JVM time zone is a GMT offset, e.g. `GMT-08:00`,
-  configure `sessionTimeZone` with offset only, e.g. `-08:00` ([jaybird#720](https://github.com/FirebirdSQL/jaybird/issues/720)) 
+  configure `sessionTimeZone` with offset only, e.g. `-08:00` ([jaybird#720](https://github.com/FirebirdSQL/jaybird/issues/720))
+- Fixed: Executing DML with a RETURNING clause containing a blob column would 
+  return the blob-id instead of the blob value ([jaybird#728](https://github.com/FirebirdSQL/jaybird/issues/728))
 
 Jaybird 4.0.8
 -------------